move tests to the file
